# good/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CRUD API for goods db",
        "title": "Good",
        "contact": {
            "name": "Developer",
            "email": "vladkolesnikofff@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/kolesnikoff17/orders-goods/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/good": {
            "get": {
                "description": "return order with given id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "good"
                ],
                "summary": "getByID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "good id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Good"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "post": {
                "description": "creates new good in repo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "good"
                ],
                "summary": "createGood",
                "parameters": [
                    {
                        "description": "name and category",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mw.GoodRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1.goodPostResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/good/{id}": {
            "put": {
                "description": "update order with given id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "good"
                ],
                "summary": "updateGood",
                "parameters": [
                    {
                        "type": "string",
                        "description": "good id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "new good data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mw.GoodRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.emptyJSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes good",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "good"
                ],
                "summary": "deleteGood",
                "parameters": [
                    {
                        "type": "string",
                        "description": "good id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.emptyJSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Good": {
            "type": "object",
            "properties": {
                "additional": {
                    "type": "object",
                    "additionalProperties": true
                },
                "category": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                }
            }
        },
        "mw.GoodRequestBody": {
            "type": "object",
            "required": [
                "category",
                "name",
                "price"
            ],
            "properties": {
                "additional": {
                    "type": "object",
                    "additionalProperties": true
                },
                "category": {
                    "type": "string",
                    "example": "Food"
                },
                "name": {
                    "type": "string",
                    "example": "Ice cream"
                },
                "price": {
                    "type": "string",
                    "example": "200"
                }
            }
        },
        "v1.emptyJSONResponse": {
            "type": "object"
        },
        "v1.goodPostResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}